move parser to bmds2
